BonTsEngine を LibISDB に置き換え
Visual Studio 2017 Update 3 以降のみ対応
Windows XP/Vista は非対応
EpgData ファイルはこれまでと互換性なし

+++ src/StatusView.cpp
@@ -1344,7 +1344,7 @@ int CStatusView::CalcTextHeight(int *pFontHeight) const
 
 int CStatusView::CalcItemHeight(int TextHeight) const
 {
-	return max(TextHeight,m_Style.IconSize.Height)+m_Style.ItemPadding.Vert();
+	return max(TextHeight,(int)m_Style.IconSize.Height)+m_Style.ItemPadding.Vert();
 }
